Integrated Project Lombok and provided annotations. Removing over 500 lines of code.
Swapped the type of ersUsersID to integer and adjusted every field which references it to call for ints. Did not do this as the AJAX/JSON level as of yet.

Adjusted html to provide HTML5 onclick events since adding event listeners to elements which weren't yet created or on screen yet was proving problematic.