Added more advanced field checking to benefit login page. Making it more generic and simultaneously more scalable for use elsewhere.
Added additional activation and animation logic so that the login button will gently pulse when active and become still and inactive when the fields are no longer valid.

Button activation is methodized to work for any element but is primarily meant for buttons.

Still have not done the register fields. Will do next.