Added additional feature to RequestViewHelper to force it to trim down the only the last segment of the URL. This fixed it so that the object should return the appropriate view no matter what computer it is run on. Allowing both of us to successfully troubleshoot.
Added additional HTML to the login page that pulls the variables from the login page automatically and updated the login() method in script.js to allow for you to manually populate a username in password in the event that debugging is ever needed in the future.

Added more logging to every class.