Ticket views populating appropriately for admins. Spotty for regular users though only one was tested with.
Set additional functionality at the end of the login method to convert the string that was the principal into a true Javascript Object so that we could reference its key value pairs.

Setup a "revealAdminPowers" function so that we could dynamically inject HTML containing the functionality for admins on the load of the dashboard if the users role is equal to 'admin'.